Type fixes
No more compiler warnings about float/double or float/int or char/int
implicit type conversions.

+++ project/engine/headers/level_manager.hpp
@@ -19,7 +19,7 @@ public:
 
 private:
     static int curlevel; 
-    static float timeStamp;
+    static double timeStamp;
     static std::string currentSound;
 };
